Update colors.md
All information that existed on this page previously now lives in Figma because Figma is the new source of truth for documenting CA brand colors. Any future updates will be reflected in Figma automatically which will eliminate the need to update duplicate documentation here in Github if/when there are changes in the future.